refactor(react-core): stop inlining @copilotkit/core in headless build, eliminate unsafe cast
The headless build previously inlined @copilotkit/core via noExternal,
creating a duplicate CopilotKitCoreReact class declaration in the .d.ts.
Since the class has private members, TypeScript treated the two declarations
as nominally distinct, forcing an `as unknown as CopilotKitContextValue`
cast in the RN provider with runtime shape validation.

@copilotkit/core has zero Node-only dependencies — the inlining was
unnecessary. By keeping it external:

- context.ts now value-exports CopilotKitCoreReact (was type-only import)
- headless.ts re-exports from ./context (externalized to package path)
- headless .d.ts references context's type → single declaration → no cast
- RN provider uses plain typed code, no as-unknown-as, no shape validation
- headless bundle is significantly smaller
