+++ packages/react-core/package.json
@@ -51,7 +51,7 @@
       "require": "./dist/v2/context.cjs"
     },
     "./v2/headless": {
-      "import": "./dist/v2/headless.js",
+      "import": "./dist/v2/headless.mjs",
       "require": "./dist/v2/headless.cjs"
     },
     "./v2/styles.css": "./dist/v2/index.css"
